Changes to Home Controller
Changes to home controller to fix errors in search. Sometimes when the player name is not found, the result variable is null and this will cause   the 500 error. Also, the case of the search player name is not the same as the case in the data for the application and this causes the result variable to be null.